GridAR-Janus-Pro/requirements.txt
accelerate==1.1.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
attrdict==2.0.1
av==17.0.0
certifi==2024.8.30
charset-normalizer==3.4.0
cmake==3.30.5
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
einops==0.8.0
exceptiongroup==1.3.1
filelock==3.16.1
flash-attn==2.7.4.post1
fonttools==4.57.0
fsspec==2024.10.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.26.2
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.4
jiter==0.13.0
kiwisolver==1.4.8
lit==18.1.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.25.2
omegaconf==2.3.0
openai==1.107.1
packaging==24.2
pillow==11.0.0
psutil==6.1.0
pycocotools==2.0.8
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.18.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
qwen-vl-utils==0.0.11
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
scipy==1.15.2
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
sympy==1.13.3
terminaltables==3.1.10
timm==0.9.16
tokenizers==0.21.1
tqdm==4.67.0
transformers @ git+https://github.com/huggingface/transformers@0ebd6651acd32c982fee265b23243b89bdb89577
triton==2.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.2.3
